service/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
PyPDF2
PyMuPDF
Pillow
requests
python-dotenv
sqlalchemy
opencv-python